Camera3: Add auto exposure lock value to metadata
Add ANDROID_CONTROL_AE_LOCK to metadata

Bug: 14059883

+++ camera/QCamera2/stack/common/cam_intf.h
@@ -460,7 +460,7 @@ typedef union {
     INCLUDE(CAM_INTF_PARM_REDEYE_REDUCTION,             int32_t,                     1);
     INCLUDE(CAM_INTF_PARM_EV,                       	int32_t,                     1);
     INCLUDE(CAM_INTF_PARM_EV_STEP,                  	cam_rational_type_t,         1);
-    INCLUDE(CAM_INTF_PARM_AEC_LOCK,                 	int32_t,                     1);
+    INCLUDE(CAM_INTF_PARM_AEC_LOCK,                 	uint8_t,                     1);
     INCLUDE(CAM_INTF_PARM_FPS_RANGE,                	cam_fps_range_t,             1);
     INCLUDE(CAM_INTF_PARM_AWB_LOCK,                 	int32_t,                     1);
     INCLUDE(CAM_INTF_PARM_EFFECT,                   	int32_t,                     1);
